+++ docker/docker.go
@@ -1,6 +1,7 @@
 package docker
 
 import (
+	"bytes"
 	"encoding/base64"
 	"fmt"
 	"os"
@@ -26,18 +27,20 @@ const (
 	BinaryNameArgument = "SPARTA_DOCKER_BINARY"
 )
 
-// BuildDockerImage creates the smallest docker image for this Golang binary
-// using the serviceName as the image name and including the supplied tags
-func BuildDockerImage(serviceName string,
+// BuildDockerImageWithFlags is an extended version of BuildDockerImage that includes
+// support for build time tags
+func BuildDockerImageWithFlags(serviceName string,
 	dockerFilepath string,
-	tags *map[string]string,
+	dockerTags map[string]string,
+	buildTags string,
+	linkFlags string,
 	logger *logrus.Logger) error {
 
 	// BEGIN DOCKER PRECONDITIONS
 	// Ensure that serviceName and tags are lowercase to make Docker happy
 	var dockerErrors []string
-	if nil != tags {
-		for eachKey, eachValue := range *tags {
+	if dockerTags != nil {
+		for eachKey, eachValue := range dockerTags {
 			if eachKey != strings.ToLower(eachKey) ||
 				eachValue != strings.ToLower(eachValue) {
 				dockerErrors = append(dockerErrors, fmt.Sprintf("--tag %s:%s MUST be lower case", eachKey, eachValue))
@@ -47,32 +50,34 @@ func BuildDockerImage(serviceName string,
 	if len(dockerErrors) > 0 {
 		return errors.Errorf("Docker build errors: %s", strings.Join(dockerErrors[:], ", "))
 	}
+	// BEGIN Informational - output the docker version...
+	dockerVersionCmd := exec.Command("docker", "-v")
+	dockerVersionCmdErr := system.RunOSCommand(dockerVersionCmd, logger)
+	if dockerVersionCmdErr != nil {
+		return errors.Wrapf(dockerVersionCmdErr, "Attempting to get docker version")
+	}
+	// END Informational - output the docker version...
+
 	// END DOCKER PRECONDITIONS
 
 	// Compile this binary for minimal Docker size
 	// https://blog.codeship.com/building-minimal-docker-containers-for-go-applications/
-	executableOutput := fmt.Sprintf("%s-%d-docker.lambda.amd64", serviceName, time.Now().UnixNano())
-	cmd := exec.Command("go",
-		"build",
-		"-a",
-		"-installsuffix",
-		"cgo",
-		"-o",
+	currentTime := time.Now().UnixNano()
+	executableOutput := fmt.Sprintf("%s-%d-docker.lambda.amd64", serviceName, currentTime)
+	buildErr := system.BuildGoBinary(serviceName,
 		executableOutput,
-		".")
-
-	cmd.Env = os.Environ()
-	cmd.Env = append(cmd.Env, "CGO_ENABLED=0", "GOOS=linux")
-	logger.WithFields(logrus.Fields{
-		"Name": executableOutput,
-	}).Info("Compiling Docker binary")
-	buildErr := system.RunOSCommand(cmd, logger)
-	if nil != buildErr {
+		false,
+		fmt.Sprintf("%d", currentTime),
+		buildTags,
+		linkFlags,
+		false,
+		logger)
+	if buildErr != nil {
 		return errors.Wrapf(buildErr, "Attempting to build Docker binary")
 	}
 	defer func() {
 		removeErr := os.Remove(executableOutput)
-		if nil != removeErr {
+		if removeErr != nil {
 			logger.WithFields(logrus.Fields{
 				"Path":  executableOutput,
 				"Error": removeErr,
@@ -95,9 +100,12 @@ func BuildDockerImage(serviceName string,
 	}
 	// Add the latest tag
 	// dockerArgs = append(dockerArgs, "--tag", fmt.Sprintf("sparta/%s:latest", serviceName))
+	logger.WithFields(logrus.Fields{
+		"Tags": dockerTags,
+	}).Info("Creating Docker image")
 
-	if nil != tags {
-		for eachKey, eachValue := range *tags {
+	if dockerTags != nil {
+		for eachKey, eachValue := range dockerTags {
 			dockerArgs = append(dockerArgs, "--tag", fmt.Sprintf("%s:%s",
 				strings.ToLower(eachKey),
 				strings.ToLower(eachValue)))
@@ -106,6 +114,21 @@ func BuildDockerImage(serviceName string,
 	dockerArgs = append(dockerArgs, ".")
 	dockerCmd := exec.Command("docker", dockerArgs...)
 	return system.RunOSCommand(dockerCmd, logger)
+}
+
+// BuildDockerImage creates the smallest docker image for this Golang binary
+// using the serviceName as the image name and including the supplied tags
+func BuildDockerImage(serviceName string,
+	dockerFilepath string,
+	tags map[string]string,
+	logger *logrus.Logger) error {
+
+	return BuildDockerImageWithFlags(serviceName,
+		dockerFilepath,
+		tags,
+		"",
+		"",
+		logger)
 }
 
 // PushDockerImageToECR pushes a local Docker image to an ECR repository
@@ -120,7 +143,7 @@ func PushDockerImageToECR(localImageTag string,
 	// 1. Get the caller identity s.t. we can get the ECR URL which includes the
 	// account name
 	stsIdentityOutput, stsIdentityErr := stsSvc.GetCallerIdentity(&sts.GetCallerIdentityInput{})
-	if nil != stsIdentityErr {
+	if stsIdentityErr != nil {
 		return "", errors.Wrapf(stsIdentityErr, "Attempting to get AWS caller identity")
 	}
 
@@ -135,7 +158,7 @@ func PushDockerImageToECR(localImageTag string,
 	// 3. Tag the local image with the ECR tag
 	dockerTagCmd := exec.Command("docker", "tag", localImageTag, ecrTagValue)
 	dockerTagCmdErr := system.RunOSCommand(dockerTagCmd, logger)
-	if nil != dockerTagCmdErr {
+	if dockerTagCmdErr != nil {
 		return "", errors.Wrapf(dockerTagCmdErr, "Attempting to tag Docker image")
 	}
 
@@ -144,17 +167,17 @@ func PushDockerImageToECR(localImageTag string,
 	var pushError error
 	dockerPushCmd := exec.Command("docker", "push", ecrTagValue)
 	pushError = system.RunOSCommand(dockerPushCmd, logger)
-	if nil != pushError {
+	if pushError != nil {
 		logger.WithFields(logrus.Fields{
 			"Error": pushError,
 		}).Info("ECR push failed - reauthorizing")
 		ecrAuthTokenResult, ecrAuthTokenResultErr := ecrSvc.GetAuthorizationToken(&ecr.GetAuthorizationTokenInput{})
-		if nil != ecrAuthTokenResultErr {
+		if ecrAuthTokenResultErr != nil {
 			pushError = ecrAuthTokenResultErr
 		} else {
 			authData := ecrAuthTokenResult.AuthorizationData[0]
 			authToken, authTokenErr := base64.StdEncoding.DecodeString(*authData.AuthorizationToken)
-			if nil != authTokenErr {
+			if authTokenErr != nil {
 				pushError = authTokenErr
 			} else {
 				authTokenString := string(authToken)
@@ -166,11 +189,13 @@ func PushDockerImageToECR(localImageTag string,
 					"login",
 					"-u",
 					authTokenParts[0],
-					"-p",
-					authTokenParts[1],
+					"--password-stdin",
 					dockerURL)
+				dockerLoginCmd.Stdout = os.Stdout
+				dockerLoginCmd.Stdin = bytes.NewReader([]byte(fmt.Sprintf("%s\n", authTokenParts[1])))
+				dockerLoginCmd.Stderr = os.Stderr
 				dockerLoginCmdErr := system.RunOSCommand(dockerLoginCmd, logger)
-				if nil != dockerLoginCmdErr {
+				if dockerLoginCmdErr != nil {
 					pushError = dockerLoginCmdErr
 				} else {
 					// Try it again...